fix: Resolve property access and initialization errors
- Fixed FileBrowserViewController tableView property access by moving it to the header.
- Corrected UIActivityIndicatorView initialization method in multiple view controllers.
- Fixed snprintf format string type mismatch in jit.m.
- Ensured all imports and dependencies are correctly handled for non-jailbroken environments.

+++ extend/jit.m
@@ -97,7 +97,7 @@ void runDebugServerCommand(int pid,
         usleep(500);
     } else {
         char attach_command[64];
-        snprintf(attach_command, sizeof(attach_command), "vAttach;%" PRIx64, pid);
+        snprintf(attach_command, sizeof(attach_command), "vAttach;%" PRIx64, (uint64_t)pid);
 
         DebugserverCommandHandle *attach_cmd = debugserver_command_new(attach_command, NULL, 0);
         if (attach_cmd == NULL) {
